Makefile: add testnet-test-write-and-read target
This works for me right now, but I don't know why I need to fork the log handle
into a reader and a writer the way I am here before using it.

If I read and write using the same handle (without making a copy of it first),
the read operations always read from the locations --write would write to
*next*, rather than the previous locations which have actually been written to.
So, there is either an off-by-one error somewhere or I am confused about how
this is supposed to work.